% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_quartet.R
\name{causal_quartet}
\alias{causal_quartet}
\title{Generate a causal quartet (constructor function)}
\usage{
causal_quartet(
  ate,
  x,
  y = NULL,
  yrange = NULL,
  yoffset = 0,
  obs = FALSE,
  varType = "random"
)
}
\arguments{
\item{ate}{A scalar.}

\item{x}{A vector.}

\item{y}{A vector.}

\item{yrange}{A vector containing the minimum and maximum x value for the y-axis.}

\item{yoffset}{A scalar indicating the baseline effect to compare to. Default=0}

\item{obs}{Boolean indicating whether quartet should show observables. Default=FALSE}

\item{varType}{The type of variation (random or systematic) to display. Default="random"}
}
\description{
Generate a causal quartet (constructor function)
}
\examples{
ate <- 0.1 
x <- c(100:120)
r <- causal_quartet(ate,x) #generate a latent causal quartet with random variation

yrange <- c(-0.5, 0.5)
r <- causal_quartet(ate,x,"yrange"=yrange) 

y <- seq(0, 1, by=0.05)
y <- y + rnorm(n=length(y), sd=0.02)
ro <- causal_quartet(ate,x,y,obs="TRUE")

s <- causal_quartet(ate,x,y,obs="TRUE", varType="systematic") 
}
