% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.causal_quartet.R
\name{print.causal_quartet}
\alias{print.causal_quartet}
\title{Print method for a causal quartet}
\usage{
\method{print}{causal_quartet}(obj)
}
\arguments{
\item{obj}{An object of class causal quartet.}
}
\description{
Print method for a causal quartet
}
\examples{
ate <- 0.5
x <- c(100:120)
r <- causal_quartet(ate,x) 
r
}
