% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quartet1_obs}
\alias{quartet1_obs}
\title{Random Quartet of potential observed data}
\format{
A data frame with 11 rows and 6 columns:
\describe{
\item{x}{values on the x-axis}
\item{y}{outcome y without the treatment effect}
\item{a}{values on the y-axis for quartet a}
\item{b}{values on the y-axis for quartet b}
\item{c}{values on the y-axis for quartet c}
\item{d}{values on the y-axis for quartet d}
...
}
}
\usage{
quartet1_obs
}
\description{
Data used to prepare Figure 3 (a-d) of Gelman, Hullman & Kennedy (2023)
}
\keyword{datasets}
