% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quartet2}
\alias{quartet2}
\title{Systematic Quartet}
\format{
A data frame with 11 rows and 6 columns:
\describe{
\item{x}{values on the x-axis}
\item{y}{only NA. Included for comparison to the observed data dataframe}
\item{a}{values on the y-axis for quartet a}
\item{b}{values on the y-axis for quartet b}
\item{c}{values on the y-axis for quartet c}
\item{d}{values on the y-axis for quartet d}
...
}
}
\usage{
quartet2
}
\description{
Data used to prepare Figure 2 of Gelman, Hullman & Kennedy (2023)
}
\keyword{datasets}
