% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quartet2_obs}
\alias{quartet2_obs}
\title{Systematic Quartet of potential observed data}
\format{
A data frame with 11 rows and 6 columns:
\describe{
\item{x}{values on the x-axis}
\item{y}{outcome y without the treatment effect}
\item{e}{values on the y-axis for quartet e}
\item{f}{values on the y-axis for quartet f}
\item{g}{values on the y-axis for quartet g}
\item{h}{values on the y-axis for quartet h}
...
}
}
\usage{
quartet2_obs
}
\description{
Data used to prepare Figure 3 (e-h) of Gelman, Hullman & Kennedy (2023)
}
\keyword{datasets}
