% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{quartet_plot_obs}
\alias{quartet_plot_obs}
\title{Plot observables consistent with an effect for a causal quartet}
\usage{
quartet_plot_obs(x, y_control, y_treat, xr, yr, ...)
}
\arguments{
\item{x}{A vector.}

\item{y_control}{A vector.}

\item{y_treat}{A vector.}

\item{xr}{A vector containing the minimum and maximum x value for the x-axis.}

\item{yr}{A vector containing the minimum and maximum y value for the y-axis.}
}
\description{
Plot observables consistent with an effect for a causal quartet
}
\examples{
x <- 4:14
y <- seq(0, 1, by=0.1)
y <- y + rnorm(n=length(y), sd=0.02)
y2 <- y + 0.1
xrange <- c(x[1], x[length(x)])
yrange <- c(-0.5, 1.5)
quartet_plot_obs(x, y, y_a_treat, xrange, yrange, main="(e)"))
}
