% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_quartet.causal_quartet.R
\name{random_quartet.causal_quartet}
\alias{random_quartet.causal_quartet}
\title{Generate y observations for a random variation causal quartet}
\usage{
\method{random_quartet}{causal_quartet}(obj)
}
\arguments{
\item{obj}{An object of class causal quartet.}
}
\description{
Generate y observations for a random variation causal quartet
}
