% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_quartet.causal_quartet.R
\name{systematic_quartet.causal_quartet}
\alias{systematic_quartet.causal_quartet}
\title{Generate y observations for a systematic variation causal quartet}
\usage{
\method{systematic_quartet}{causal_quartet}(obj)
}
\arguments{
\item{obj}{An object of class causal quartet.}
}
\description{
Generate y observations for a systematic variation causal quartet
}
