% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_quartet.R
\name{causal_quartet}
\alias{causal_quartet}
\title{Generate a causal quartet}
\usage{
causal_quartet(
  ate,
  x,
  y = NULL,
  yrange = NULL,
  yoffset = 0,
  obs = FALSE,
  vartype = "random",
  matchlq = NULL
)
}
\arguments{
\item{ate}{A scalar}

\item{x}{A vector}

\item{y}{A vector}

\item{yrange}{A vector containing the minimum and maximum x value for the y-axis. Default=NULL}

\item{yoffset}{A scalar indicating the baseline effect to compare to. Default=0}

\item{obs}{Boolean indicating whether quartet should show observables. Default=FALSE}

\item{vartype}{The type of variation (random or systematic) to display. Default="random"}

\item{matchlq}{A causal quartet object a latent quartet, for use in generating a corresponding observables quartet. Default=NULL}
}
\value{
A list object of class causal_quartet. The object has the following elements:
\itemize{
\item{"x"}{ Contains user-provided x observations}
\item{"y"}{ Contains user-provided control y observations for an observables quartet}
\item{"a"}{ Contains y values for subplot a. If an observables quartet, these are y values for the treatment condition}
\item{"b"}{ Contains y values for subplot b. If an observables quartet, these are y values for the treatment condition}
\item{"c"}{ Contains y values for subplot c. If an observables quartet, these are y values for the treatment condition}
\item{"d"}{ Contains y values for subplot d. If an observables quartet, these are y values for the treatment condition}
}
The object has the following attributes:
\itemize{
\item{"ate"}{ The user-provided average treatment effect}
\item{"yoffset"}{ The baseline y value indicating no effect in a latent quartet}
\item{"xrange"}{ The x-axis range (min and max)}
\item{"yrange"}{ The y-axis range (min and max)}
\item{"yrange_given"}{ Indicates whether the user specified a y-axis range}
\item{"space"}{ Whether the quartet plots latent effects (latent) or hypothetical observables (observables)}
\item{"variation"}{ Whether the variation is random or systematic}
}
}
\description{
Generate a causal quartet
}
\examples{
ate <- 0.1 
x <- c(100:120)
r <- causal_quartet(ate,x) #generate a latent causal quartet with random variation

yrange <- c(-0.5, 0.5)
r <- causal_quartet(ate,x,"yrange"=yrange) 

y <- seq(0, 1, by=0.05)
y <- y + rnorm(n=length(y), sd=0.02)
ro <- causal_quartet(ate,x,y,obs="TRUE")

s <- causal_quartet(ate,x,y,obs="TRUE", vartype="systematic") 
}
